+++ app/javascript/components/sidebar/active-layers-indicator/index.js
@@ -1,8 +1,0 @@
-import { connect } from 'react-redux';
-
-// component
-import ActiveLayersIndicator from './component';
-
-export default connect(
-  state => ({ activeLayers: state.fspMaps.layers.selectedLayers })
-)(ActiveLayersIndicator);